clean up layout
