+++ src/pages/_app.tsx
@@ -5,11 +5,21 @@ import type { AppType } from 'next/dist/shared/lib/utils';
 import superjson from 'superjson';
 import '../styles/globals.css';
 import { connectZustandStateToReduxDevtools } from 'store/reduxDevToolsConfig';
+import Head from 'next/head';
 
 connectZustandStateToReduxDevtools();
 
 const MyApp: AppType = ({ Component, pageProps }) => {
-  return <Component {...pageProps} />;
+  return (
+    <>
+      <Head>
+        <title>Eco Admin App</title>
+        <meta name="description" content="Generated by create-t3-app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <Component {...pageProps} />
+    </>
+  );
 };
 
 const getBaseUrl = () => {
