+++ src/pages/index.tsx
@@ -1,69 +1,10 @@
 import type { NextPage } from 'next';
-import Head from 'next/head';
-import { useEffect, useRef } from 'react';
-import { RecentUsersSection, Spinner, SelectedUserSection } from 'components';
-import { trpc } from 'utils/trpc';
-import { useUsers } from 'store';
+import Users from './users';
 
 const Home: NextPage = () => {
-  const { setRecentUsers, recentUsers, selectedUser, setLoading } = useUsers();
-  const { data: usersData, isLoading: usersLoading } = trpc.useQuery([
-    'example.recentUsers',
-  ]);
-
-  useEffect(() => {
-    if (usersData && usersData.users) {
-      setRecentUsers(usersData.users);
-    }
-  }, [setRecentUsers, usersData]);
-
-  useEffect(() => {
-    setLoading(usersLoading);
-  }, [setLoading, usersLoading]);
-
-  const selectedUserRef = useRef(null);
-
-  const scrollToBottom = () => {
-    // @ts-ignore:next-line
-    selectedUserRef?.current?.scrollIntoView({ behavior: 'smooth' });
-  };
-
-  useEffect(() => {
-    if (selectedUser) {
-      scrollToBottom();
-    }
-  }, [selectedUser]);
-
   return (
     <>
-      <Head>
-        <title>Eco Admin App</title>
-        <meta name="description" content="Generated by create-t3-app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      {usersLoading && (
-        <div className="flex items-center justify-center h-screen">
-          <Spinner />
-        </div>
-      )}
-
-      {!!recentUsers && !usersLoading && (
-        <div className="bg-gray-100 h-full w-full py-10">
-          <div className="max-w-10xl mx-auto px-4 sm:px-6 lg:px-4">
-            <div className="max-w-7xl mx-auto">
-              <div className="p-4 bg-white">
-                <RecentUsersSection users={recentUsers} />
-              </div>
-
-              {selectedUser && (
-                <div ref={selectedUserRef}>
-                  <SelectedUserSection user={selectedUser} />
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
+      <Users />
     </>
   );
 };
